Created a custom layout component
Created such a component to keep the basic components that should be on each page together (for example, the dark mode setting).
