+++ src/pages/index.js
@@ -5,33 +5,18 @@
  */
 
 import React from "react"
-import Navbar from "../components/Navbar"
 import Landing from "../sections/Landing"
 import About from "../sections/About"
 import Team from "../sections/Team"
-import { Helmet } from "react-helmet"
-import useDarkModeState from "../libs/useDarkModeState"
-
-// import Download from "../sections/Download"
+import Layout from "../components/Layout"
 
 const IndexPage = () => {
-
-  const isDark = useDarkModeState();
-
   return (
-    <main className={isDark ? "dark" : "light"}>
-      <Helmet htmlAttributes={{ lang: 'en' }}>
-        <title>Skyline - Nintendo Switch Emulator</title>
-        <meta name="description" content="Skyline is an open source and experimental emulator that runs on ARMv8 Android™ devices and emulates the functionality of a Nintendo Switch™ system." />
-        <meta name="keywords" content="Emulator, Nintendo Switch™, Skyline, Android" />
-        <meta name="author" content="Skyline Team" />
-      </Helmet>
-      <Navbar />
+    <Layout title="Skyline - Nintendo Switch Emulator">
       <Landing />
       <About/>
       <Team />
-      {/* <Download /> */}
-    </main> 
+    </Layout> 
   )
 }
